Move service interfaces into the main package
