+++ src/main/java/com/coinpaprika/apiclient/MoversRange.kt
@@ -1,0 +1,5 @@
+package com.coinpaprika.apiclient
+
+enum class MoversRange {
+    TOP200, TOP300, ALL
+}
