--- src/main/java/com/coinpaprika/apiclient/repository/ranking/MoversRange.kt
@@ -1,5 +1,0 @@
-package com.coinpaprika.apiclient.repository.ranking
-
-enum class MoversRange {
-    TOP200, TOP300, ALL
-}
